requirements.txt
tf-nightly==2.3.0.dev20200613
tensorflow_datasets
numpy
matplotlib
